# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Catering Business API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Admin menu endpoints
...
  /plan:
    get:
      operationId: getPlan
      tags: [orders]
      summary: Get saved plan (wishlist)
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plan with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    post:
      operationId: addToPlan
      tags: [orders]
      summary: Add item to plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToPlanBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plan/{itemId}:
    delete:
      operationId: removeFromPlan
      tags: [orders]
      summary: Remove item from plan
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/orders/{id}:
    put:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # AI Chat
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        servingSize:
          type: integer
          description: Number of people one unit serves
        unit:
          type: string
          description: e.g. "tray", "dozen", "per person"
        imageUrl:
          type: string
          nullable: true
        allergens:
          type: array
          items:
            type: string
        available:
          type: boolean
        prepTime:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - category
        - price
        - servingSize
        - unit
        - allergens
        - available
        - createdAt
...
    PlanItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItem:
          $ref: "#/components/schemas/MenuItem"
      required:
        - id
        - menuItemId
        - menuItem
    Plan:
      type: object
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlanItem"
      required:
        - sessionId
        - items
    AddToPlanBody:
      type: object
      properties:
        sessionId:
          type: string
        menuItemId:
          type: integer
      required:
        - sessionId
        - menuItemId
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required:
        - id
        - menuItemId
        - menuItemName
        - price
        - quantity
    Order:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        eventDate:
          type: string
          format: date
          nullable: true
        eventType:
          type: string
          nullable: true
        guestCount:
          type: integer
          nullable: true
        serviceStyle:
          type: string
          nullable: true
        deliveryNotes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, preparing, delivered, cancelled]
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - customerName
        - customerEmail
        - status
        - total
        - items
        - createdAt
    CreateOrderBody:
      type: object
      properties:
        sessionId:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        eventDate:
          type: string
          format: date
          nullable: true
        eventType:
          type: string
          nullable: true
        guestCount:
          type: integer
          nullable: true
        serviceStyle:
          type: string
          nullable: true
        deliveryNotes:
          type: string
          nullable: true
      required:
        - sessionId
        - customerName
        - customerEmail
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, delivered, cancelled]
      required:
        - status
...