# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catering Business API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  sessionId: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  eventDate?: Date | null;
  eventType?: string | null;
  guestCount?: number | null;
  serviceStyle?: string | null;
  deliveryNotes?: string | null;
}