# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catering Business API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  sessionId: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  eventDate?: Date | null;
  eventType?: string | null;
  guestCount?: number | null;
  serviceStyle?: string | null;
  deliveryNotes?: string | null;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  createdAt: Date;
}