# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catering Business API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanItem } from "./planItem";

export interface Plan {
  sessionId: string;
  items: PlanItem[];
}