# lib/api-zod/src/generated/types/planItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catering Business API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItem } from "./menuItem";

export interface PlanItem {
  id: number;
  menuItemId: number;
  menuItem: MenuItem;
}